+++ lib/src/widget/forme_field_builder.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:forme/forme.dart';
 
 class FormeFieldBuilder<E extends FormeModel> extends StatelessWidget {